requirements.txt
# Core app libraries
Flask
requests
slack_sdk
openai

# OpenTelemetry Core
opentelemetry-distro
opentelemetry-api
opentelemetry-sdk

# OpenTelemetry Instrumentation
opentelemetry-instrumentation-flask
opentelemetry-instrumentation-requests
opentelemetry-instrumentation-threading
opentelemetry-instrumentation-logging

# OpenTelemetry Exporters
opentelemetry-exporter-otlp-proto-http
opentelemetry-exporter-prometheus

# OpenTelemetry Semantic Conventions
opentelemetry-semantic-conventions